Implemented CRUD operations in UserController.cs
The `Microsoft.CodeDom.Providers.DotNetCompilerPlatform` reference was added and then removed from the project file. The closing `</Project>` tag was also removed and then added back in due to a potential formatting or version control issue.

In `UserController.cs`, CRUD operations for `User` objects were implemented. The `Index` method now returns a view of the `userlist`. The `Details` method finds a user based on the provided ID and returns a view of that user or a `HttpNotFoundResult` if no user is found. The `Create` methods handle form display and submission for creating a new user. The `Edit` methods handle form display and submission for editing an existing user. The `Delete` methods handle form display and submission for deleting a user.