Implemented CRUD operations in UserController.cs
Implemented the `Index`, `Details`, `Create`, `Edit`, and `Delete` methods in `UserController.cs` to handle user data. These methods now contain logic for retrieving, adding, updating, and deleting users.

The `CRUD application 2.csproj` file was temporarily modified to include a reference to `Microsoft.CodeDom.Providers.DotNetCompilerPlatform`, which was later removed. The closing `Project` tag was also moved to a new line.

The `Web.config` file was updated to change the `oldVersion` attribute of the `bindingRedirect` tag for `WebGrease` to `0.0.0.0-1.6.5135.21930`, allowing any version from `0.0.0.0` to `1.6.5135.21930` to be redirected to use version `1.6.5135.21930`.