Implemented UserController methods and updated WebGrease version range
Implemented `Index`, `Details`, `Create`, `Edit`, and `Delete` methods in `UserController.cs`, adding significant functionality to the `UserController` class. In `Web.config`, updated the `oldVersion` attribute of the `bindingRedirect` tag for `WebGrease` to include a wider range of versions. Minor changes in `CRUD application 2.csproj` include adding and removing a reference to `Microsoft.CodeDom.Providers.DotNetCompilerPlatform` and a formatting correction in the closing `Project` tag.