Updated deployment settings and added GitHub Actions workflow
Updated the `.csproj` and `deploy.parameters.json` files to include new files and modify the `webAppName` parameter. Added a new `publishsettings.json` file with the publish details for the `crud-app-dom` application. Also, a new `main.yml` file has been added under `.github\workflows` to define a GitHub Actions workflow that triggers on push to the master branch, sets up a .NET Core environment, builds and publishes the application, and deploys it to an Azure Web App.